requirements.txt
Pillow==3.3.1
python-aalib==0.3.2
requests==2.11.1
wikipedia==1.4.0
